Changed the return type of strtoull() from 'unsigned long long' (which is not valid C89) to 'uint64_t' (declared in <net-snmp/net-snmp-config.h>). This change makes the Net-SNMP source code compile again with MSVC 6.0.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@17869 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/include/net-snmp/library/system.h
@@ -137,7 +137,7 @@ SOFTWARE.
     unsigned long   strtoul(const char *, char **, int);
 #endif
 #ifndef HAVE_STRTOULL
-    unsigned long long strtoull(const char *, char **, int);
+    uint64_t        strtoull(const char *, char **, int);
 #endif
 #ifndef HAVE_STRTOK_R
     char           *strtok_r(char *, const char *, char **);
